Update fi.po with translations
Added translations for the shortened compass rose directions. However I would note, there is not really standard for this in Finnish language use and these are little contrived. We have separate words for the main intermediates, but these don't shorten usually and even less for the octal intermediates, which would take combining intermediate word with cardinal. Which is done, but shortens even worse.

So i did combining the cardinals. Which as grammatical Finnish and has been done before given some examples I have seen on looking for it, but is rarely used commonly in normal Finnish parlance.